backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
boto3==1.40.64
botocore==1.40.64
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-csp==4.0
django-environ==0.12.0
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
imagedominantcolor==1.0.1
ImageHash==4.3.2
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
numpy==2.4.2
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyWavelets==1.9.0
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
s3transfer==0.14.0
scipy==1.17.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
videohash==3.0.1
vine==5.1.0
wcwidth==0.2.14
yt-dlp==2026.2.21
